+++ src/private/mod.rs
@@ -1,4 +1,19 @@
 //! Private helpers
 
+// /// Private trait to implement sealing for [`SyncNow`]
+// #[cfg(feature = "std")]
+// pub trait SyncNowSealed {}
+
+/// Private trait to implement sealing for [`ErasedSleep`]
+pub trait ErasedSleepSealed<D> {}
+
+/// Private trait to implement sealing for [`StdNow`]
+#[cfg(feature = "std")]
+pub trait StdNowSealed {}
+
+/// Private trait to implement sealing for [`ChronoNow`]
 #[cfg(feature = "chrono04")]
-pub mod chrono;
+pub trait ChronoNowSealed {}
+
+/// Private trait to implement sealing for [`NowExt`]
+pub trait NowExtSealed {}