+++ src/std_system.rs
@@ -1,15 +1,17 @@
-use crate::Clock;
+use crate::impl_now;
 use ::std::time::Instant;
 
 /// Standard library clock using the system time.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct SystemStdClock;
 
-impl Clock for SystemStdClock {
-  type Instant = Instant;
+impl_now! {
+  impl Now for SystemStdClock {
+    type Instant = Instant;
 
-  fn now(&self) -> Self::Instant {
-    Instant::now()
+    fn now(&this)-> Self::Instant {
+      Instant::now()
+    }
   }
 }
 
@@ -19,7 +21,7 @@ impl Clock for SystemStdClock {
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::StdClock;
+  use crate::{Now, NowOnce, StdNow};
   use ::std::time::Duration;
   const ONE_YEAR: Duration = Duration::new(365 * 24 * 60 * 60, 0);
 
@@ -30,19 +32,19 @@ mod tests {
     use_std_clock(&clock);
   }
 
-  fn use_clock<TyClock>(clock: &TyClock)
+  fn use_clock<TyNow>(clock: &TyNow)
   where
-    TyClock: Clock<Instant = Instant>,
+    TyNow: Now<Instant = Instant>,
   {
     let one_year_ago = Instant::now() - ONE_YEAR;
-    let now = clock.now();
+    let now = clock.now_once();
     assert!(now > one_year_ago);
     use_std_clock(clock);
   }
 
   fn use_std_clock<TyClock>(clock: &TyClock)
   where
-    TyClock: StdClock,
+    TyClock: StdNow,
   {
     let one_year_ago = Instant::now() - ONE_YEAR;
     let now = clock.now_std();